backend/bnShopApi/requirements.txt
asgiref==3.5.2
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
distlib==0.3.5
dj-database-url==1.0.0
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
facebook-sdk==3.1.0
filelock==3.7.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.21.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
platformdirs==2.5.2
protobuf==4.21.9
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.16.3
whitenoise==6.2.0
